ingest/requirements.txt
# Data Ingestion Module Requirements
requests>=2.28.0
aiohttp>=3.8.0
feedparser>=6.0.0
beautifulsoup4>=4.11.0
selenium>=4.5.0
tweepy>=4.12.0
praw>=7.6.0
redis>=4.3.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
structlog>=22.1.0
click>=8.1.0
tqdm>=4.64.0